requirements.txt
# ── AI / Script ─────────────────────────────────────────────────
requests>=2.31.0
groq>=0.4.0                  # Groq API client (optional, requests fallback included)

# ── Voiceover ────────────────────────────────────────────────────
edge-tts>=6.1.9              # Microsoft Edge TTS — free, no API key

# ── Video editing ────────────────────────────────────────────────
moviepy>=1.0.3
openai-whisper>=20231117     # Auto-captions (runs locally, free)
numpy>=1.24.0

# ── YouTube upload ───────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# ── Utilities ────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
